flask/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Werkzeug==2.2.2
Flask-WTF==1.0.1
bcrypt==4.1.2
email-validator==2.0.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
SQLAlchemy-serializer==1.4.1
Flask-Session==0.6.0
pyjwt==2.8.0
Authlib==1.3.0
requests==2.31.0